Verify email landing page db edit
Currently the verify email token document is not being deleted
Strict mode calls the verify email function twice
On first call the user info is correct but on 2nd call,
it is gone as the verify email token doc is deleted.
How to I make this work whilst keeping the tokens collection clean?